rename Admin nav, add font fix for android
