+++ app/navigators/AdminNavigator.js
