+++ app/navigators/index.js
@@ -1,0 +1,3 @@
+export { default as AdminNavigator } from './AdminNavigator';
+export { default as EmployeeNavigator } from './EmployeeNavigator';
+export { default as RefugeeNavigator } from './RefugeeNavigator';
