+++ package.json
@@ -8,6 +8,8 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^9.0.0",
+    "@types/react": "^16.7.22",
+    "@types/react-navigation": "^3.0.1",
     "expo": "^32.0.0",
     "native-base": "^2.10.0",
     "react": "16.5.0",
